+++ pom.xml
@@ -18,11 +18,11 @@
   xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
   <modelVersion>4.0.0</modelVersion>
 
-  <groupId>org.saladframework</groupId>
-  <artifactId>salad-orm</artifactId>
+  <groupId>org.simple.framework</groupId>
+  <artifactId>simple-orm</artifactId>
   <version>1.0-SNAPSHOT</version>
 
-  <name>salad-orm</name>
+  <name>simple-orm</name>
 
   <developers>
     <developer>
