--- src/main/java/org/saladframework/dao/KeyValue.java
@@ -1,55 +1,0 @@
-/*
- *    Licensed under the Apache License, Version 2.0 (the "License");
- *    you may not use this file except in compliance with the License.
- *    You may obtain a copy of the License at
- *
- *        http://www.apache.org/licenses/LICENSE-2.0
- *
- *    Unless required by applicable law or agreed to in writing, software
- *    distributed under the License is distributed on an "AS IS" BASIS,
- *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *    See the License for the specific language governing permissions and
- *    limitations under the License.
- */
-
-package org.saladframework.dao;
-
-/**
- * KeyValue
- *
- * @author CuiShiFeng
- * @date
- */
-public class KeyValue<K, V> {
-
-    private K key;
-
-    private V value;
-
-
-    public KeyValue(K key, V value) {
-        super();
-        this.key = key;
-        this.value = value;
-    }
-
-    public KeyValue() {
-
-    }
-
-    public K getKey() {
-        return key;
-    }
-
-    public void setKey(K key) {
-        this.key = key;
-    }
-
-    public V getValue() {
-        return value;
-    }
-
-    public void setValue(V value) {
-        this.value = value;
-    }
-}
