+++ src/main/java/org/simple/framework/dao/KeyValueParis.java
@@ -1,0 +1,40 @@
+/*
+ *    Licensed under the Apache License, Version 2.0 (the "License");
+ *    you may not use this file except in compliance with the License.
+ *    You may obtain a copy of the License at
+ *
+ *        http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *    Unless required by applicable law or agreed to in writing, software
+ *    distributed under the License is distributed on an "AS IS" BASIS,
+ *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *    See the License for the specific language governing permissions and
+ *    limitations under the License.
+ */
+
+package org.simple.framework.dao;
+
+import java.util.HashMap;
+import java.util.Map;
+
+/**
+ * KeyValueParis
+ *
+ * @author CuiShiFeng
+ * @date
+ */
+public class KeyValueParis {
+
+    private Map<String, Object> kvPairs = new HashMap<String, Object>();
+
+    public KeyValueParis add(String key, Object value) {
+        kvPairs.put(key, value);
+
+        return this;
+    }
+
+    public Map<String, Object> getKvPairs() {
+        return kvPairs;
+    }
+
+}
