+++ src/main/java/org/simple/framework/dao/UnExpectedException.java
@@ -1,0 +1,46 @@
+/*
+ *    Licensed under the Apache License, Version 2.0 (the "License");
+ *    you may not use this file except in compliance with the License.
+ *    You may obtain a copy of the License at
+ *
+ *        http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *    Unless required by applicable law or agreed to in writing, software
+ *    distributed under the License is distributed on an "AS IS" BASIS,
+ *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *    See the License for the specific language governing permissions and
+ *    limitations under the License.
+ */
+
+package org.simple.framework.dao;
+
+/**
+ * 不期望的异常
+ *
+ * @author LiuKeFeng
+ * @date 2017年4月10日
+ */
+public class UnExpectedException extends RuntimeException {
+    private static final long serialVersionUID = 6409066485217841653L;
+
+    public UnExpectedException() {
+        super();
+    }
+
+    public UnExpectedException(String message, Throwable cause, boolean enableSuppression, boolean writableStackTrace) {
+        super(message, cause, enableSuppression, writableStackTrace);
+    }
+
+    public UnExpectedException(String message, Throwable cause) {
+        super(message, cause);
+    }
+
+    public UnExpectedException(String message) {
+        super(message);
+    }
+
+    public UnExpectedException(Throwable cause) {
+        super(cause);
+    }
+
+}
