+++ src/main/java/org/simple/framework/dao/pool/ConnectionPool.java
@@ -1,0 +1,25 @@
+package org.simple.framework.dao.pool;
+
+import java.sql.Connection;
+import java.sql.SQLException;
+
+/**
+ * connection pool.
+ * @author simple <cuishifeng0207@163.com>
+ * Created on 2022-04-10
+ */
+public interface ConnectionPool {
+
+    /**
+     * get connection
+     * @return
+     */
+    Connection getConnection();
+
+    /**
+     * release connection
+     * @param conn
+     * @throws SQLException
+     */
+    void release(Connection conn) throws SQLException;
+}
