+++ src/main/java/org/simple/framework/dao/util/OutSQL.java
@@ -1,0 +1,41 @@
+/*
+ *    Licensed under the Apache License, Version 2.0 (the "License");
+ *    you may not use this file except in compliance with the License.
+ *    You may obtain a copy of the License at
+ *
+ *        http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *    Unless required by applicable law or agreed to in writing, software
+ *    distributed under the License is distributed on an "AS IS" BASIS,
+ *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *    See the License for the specific language governing permissions and
+ *    limitations under the License.
+ */
+
+package org.simple.framework.dao.util;
+
+/**
+ * @author CuiShiFeng
+ * @date
+ */
+public class OutSQL {
+
+    private String sql;
+
+    public OutSQL() {
+
+    }
+
+    public OutSQL(String sql) {
+        this.sql = sql;
+    }
+
+    public void setSql(String sql) {
+        this.sql = sql;
+    }
+
+    public String getSql() {
+        return sql;
+    }
+
+}
